[ENH] Major updates to pypmi.bids
Trying to make it a real module! Modifies the conversion process so that
it (mostly) works, for now. Still needs some updates to ensure that the
outputs are valid BIDS datasets, but working on it. Once a new
`heudiconv` release is made we should be able to get rid of the nasty
custom_callable in the heuristic.py for fixing the multiecho PDT2
scans...