[TEST] Check only for downloaded files in test
Somehow the travis cache was including an extra file? Or something. So
the listdir() call to the studydata directory had one more file than was
expected, resulting in an assertion error right at pytest setup.

Now we explicitly check for each file that we wanted to download: was it
downloaded? If not, fail. Might be good to move this check into the
fetcher_*() functions themselves?